+++ raytracer/main/color.h
@@ -7,22 +7,22 @@
 struct Color
 {
     Color() : red{0}, green{0}, blue{0} {}
-    Color(double red, double green, double blue);
+    Color(float red, float green, float blue);
 
     Color operator +(const Color& color) const;
     Color operator -(const Color& color) const;
-    Color operator *(const double& multiplier) const;
+    Color operator *(const float& multiplier) const;
     Color operator *(const Color& color) const;
-    Color operator /(const double& divider) const;
+    Color operator /(const float& divider) const;
     bool operator ==(const Color& color) const;
 
     friend std::ostream &operator<<(std::ostream &os, const Color& c) {
         return os << "(" << c.red << ", " << c.green << ", " << c.blue << ")";
     }
 
-    double red;
-    double green;
-    double blue;
+    float red;
+    float green;
+    float blue;
 };
 
 struct Pixel
